requirements/local.txt
# Local development dependencies go here
-r base.txt
django-debug-toolbar==0.11
Sphinx==1.2.2
mock==1.0.1
